Implements polymorphic comments
Adds support for polymorphic comments, allowing comments to be associated with different models.

This introduces a new `commented` morph relation to the `Comment` model, enabling comments to be linked to various entities.

The `User` model is updated to be commentable using the `Commentable` trait.
A new `HasComments` trait is introduced to handle the commented side of the relation.