Refactors module route and provider stubs
Updates the module route provider stub to use a more streamlined approach for defining routes, utilizing the Locale facade for setting the locale.

Simplifies the module provider stub by removing redundant code and updating the registration of translations and views.

The config stub is also updated to remove the repositories array.